+++ catshflow/history/constant/history_constant.py
@@ -1,0 +1,5 @@
+import os
+
+current_path = os.path.dirname(os.path.abspath(__file__))
+BASE_PATH = os.path.sep.join(current_path.split(os.path.sep)[:-2] + ["temp"])
+MOVEMENTS_FILE = os.path.join(BASE_PATH, "movement.json")