+++ catshflow/login/exceptions/login_exceptions.py
@@ -1,0 +1,3 @@
+class NotFound(Exception):
+    def __init__(self, *args):
+        super().__init__(*args)